requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
attrdict==2.0.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.16.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.9.0
Flask==3.0.2
fonttools==4.49.0
freetype-py==2.4.0
fsspec==2023.4.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
grpcio==1.62.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipython==8.18.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.39.1
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nbformat==5.5.0
nest-asyncio==1.6.0
networkx==2.2
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.16.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybullet==3.2.5
pybulletX==0.4.1
pycollada==0.6
pyglet==2.0.12
Pygments==2.17.2
PyOpenGL==3.1.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.18.0
rsa==4.9
sb3-contrib==2.0.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
Shimmy==1.3.0
six==1.16.0
smmap==5.0.1
stable-baselines3==2.0.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorboard==2.12.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.3.0
tqdm==4.66.2
traitlets==5.14.1
trimesh==4.1.7
triton==2.2.0
typing_extensions==4.8.0
tzdata==2024.1
urdfpy==0.0.22
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
zipp==3.17.0
